# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/subcategories": {
            "get": {
                "tags": [
                    "Subcategories"
                ],
                "summary": "Get all subcategories",
                "operationId": "fb91228103bb68bf2244f7c1cb79769a",
                "responses": {
                    "200": {
                        "description": "List of subcategories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "position": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "subcat_name": {
                                                "type": "string",
                                                "example": "Subcat name"
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "categories": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "position": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "category_name": {
                                                        "type": "string",
                                                        "example": "Category name"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Subcategories"
                ],
                "summary": "Create a new subcategory",
                "operationId": "63c527b456c59e19aba9e022502feec1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "subcat_name",
                                    "position",
                                    "category_id"
                                ],
                                "properties": {
                                    "subcat_name": {
                                        "type": "string",
                                        "example": "New Subcategory"
                                    },
                                    "position": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subcategory created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subcategory created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{categoryId}/subcategories": {
            "get": {
                "tags": [
                    "Subcategories"
                ],
                "summary": "Get all subcategories by category",
                "description": "Returns a list of subcategories for a specific category",
                "operationId": "979a8241578dfca23be533c4f09042a5",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of subcategories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "position": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "subcat_name": {
                                                "type": "string",
                                                "example": "Subcat name"
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "categories": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "position": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "category_name": {
                                                        "type": "string",
                                                        "example": "Category name"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/subcategories/{id}": {
            "get": {
                "tags": [
                    "Subcategories"
                ],
                "summary": "Get a subcategory by ID",
                "operationId": "1119ebac20e0aa36565466bdd484c46a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the subcategory to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "position": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "subcat_name": {
                                            "type": "string",
                                            "example": "Subcat name"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subcategory not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subcategory not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Subcategories"
                ],
                "summary": "Update a subcategory",
                "operationId": "e7302a6ebda00a1f621bccae93b1b80c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the subcategory to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "subcat_name",
                                    "position",
                                    "category_id"
                                ],
                                "properties": {
                                    "subcat_name": {
                                        "type": "string",
                                        "example": "Updated Subcategory"
                                    },
                                    "position": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subcategory updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subcategory updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subcategories"
                ],
                "summary": "Delete a subcategory",
                "operationId": "3f97052dfd3d10382a706e0998d69092",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the subcategory to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subcategory deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subcategory deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subcategory not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subcategory not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}